0.17.37

+++ Source/frmMassCompare.cs
@@ -78,7 +78,7 @@ namespace UniversalPatcher
                 if (peekPCM.tableDatas[ind].Rows == 1 && peekPCM.tableDatas[ind].Columns == 1)
                 {
                     double curVal = getValue(peekPCM.buf, (uint)(peekPCM.tableDatas[ind].addrInt + peekPCM.tableDatas[ind].Offset), peekPCM.tableDatas[ind],0,peekPCM);
-                    UInt64 rawVal = getRawValue(peekPCM.buf,(uint)(peekPCM.tableDatas[ind].addrInt + peekPCM.tableDatas[ind].Offset), peekPCM.tableDatas[ind],0);
+                    UInt64 rawVal = (UInt64) getRawValue(peekPCM.buf,(uint)(peekPCM.tableDatas[ind].addrInt + peekPCM.tableDatas[ind].Offset), peekPCM.tableDatas[ind],0);
                     string valTxt = curVal.ToString();
                     string unitTxt = " " + peekPCM.tableDatas[ind].Units;
                     string maskTxt = "";
